previous push didn't suffice, made some other chages to code

+++ api/create_user.php
@@ -19,29 +19,40 @@ $user = new User($db);
  
 // get posted data
 $data = json_decode(file_get_contents("php://input"));
- 
+
 // set product property values
+$user->email = $data->email;
 $user->firstname = $data->firstname;
 $user->lastname = $data->lastname;
-$user->email = $data->email;
 $user->password = $data->password;
- 
-// create the user
-if($user->create()){
- 
-    // set response code
-    http_response_code(200);
- 
-    // display message: user was created
-    echo json_encode(array("message" => "User was created."));
-}
- 
-// message if unable to create user
-else{
- 
+$user->contact_number = $data->contact_number;
+$user->address = $data->address;
+$user->access_level = $data->access_level;
+$user->status = 0;
+//check if email already exists
+if($user->emailExists()){
     // set response code
     http_response_code(400);
- 
+
     // display message: unable to create user
-    echo json_encode(array("message" => "Unable to create user."));
+    echo json_encode(array("message" => "Error! Email already exist"));
+}else{
+    // create the user
+    if($user->create()){
+
+        // set response code
+        http_response_code(200);
+    
+        // display message: user was created
+        echo json_encode(array("message" => "User was created."));
+    }
+    // message if unable to create user
+    else{
+    
+        // set response code
+        http_response_code(400);
+    
+        // display message: unable to create user
+        echo json_encode(array("message" => "Unable to create user."));
+    }
 }
